build(libcore): show libcore sha256 About #1

+++ libcore/init.sh
@@ -1,9 +1,13 @@
 #!/bin/bash
+set -e
+
+[[ -e "inited" ]] && exit 0
 
 # Install gomobile
 if [ ! -f "$GOPATH/bin/gomobile" ]; then
-    go install golang.org/x/mobile/cmd/gomobile@latest
-    go install golang.org/x/mobile/cmd/gobind@latest
+	go install golang.org/x/mobile/cmd/gomobile@latest
+	go install golang.org/x/mobile/cmd/gobind@latest
 fi
 
 gomobile init
+touch inited
